Convert to GSON

+++ app/src/main/java/uk/co/maddwarf/randomdungeongeneratorpremium/model/Monster.kt
@@ -3,7 +3,7 @@ package uk.co.maddwarf.randomdungeongeneratorpremium.model
 data class Monster (
     val name: String,
     val xp: Int = 0,
-    val cr: Int = 0,
+    val cr: Float = 0f,
     val weight: Int = 1
 
 )
